Move api::MediaDisplayInfo and api::MediaDeviceInfo into separate modules
